Updates to WCAG 2.0 checks
Don't report error if there is a onclick or onkeypress attribute on a
tag if there is a focusable item inside a block tag - WCAG_2.0-F42.
Add <button> tag to list of tags that are allowed event handlers -
WCAG_2.0-F42.
Report WCAG_2.0-G131 for missing label and legend content.
Check for styled text (e.g. bold) that looks like a heading -
WCAG_2.0-F2.
Check for missing table headers at the end of the table. This allows for
header id definitions to come after a reference - WCAG_2.0-H43.