Add routines to perform content checks
Check for leading or trailing whitespace in field values
Check for identical values for all cells in a column
Check for next line in multi-line cell before checking for a new list item.
Stop recording errors if maximum number allowed is reached.
Check for inconsistencies in spacing, punctuation and pluralization in text cells